feat: 내 정보 조회 폴더 이동 및 로직 수정
feat: 내 정보 조회 폴더 이동 로직 수정
